+++ src/page/cardAdd/index.jsx
@@ -1,0 +1,181 @@
+import { useState, useCallback, useEffect } from 'react';
+
+import FormInput from '../../components/common/FormInput';
+import CardPreview from '../../components/CardPreview';
+import Modal from '../../components/common/Modal';
+import CardCompany from '../../components/CardCompany';
+import Header from '../../components/common/Header';
+import Button from '../../components/common/Button';
+
+import { ReactComponent as PrevIcon } from '../../assets/prev_icon.svg';
+import Tooltip from '../../components/common/Tooltip';
+import { CRYPTO_STRING, INPUT_MAX_LENGTH } from '../../constants';
+import {
+  cardNumberInputInfoList,
+  expiryDateInputInfoList,
+  cardOwnerNameInputInfoList,
+  privacyCodeInputInfoList,
+  cardPasswordInputInfoList,
+  cardCompanyList,
+} from './data';
+
+import { validator, checkFullFilled } from './validator';
+import { isObject } from '../../utils';
+
+const initialCardInfo = {
+  company: '',
+  number: {
+    first: '',
+    second: '',
+    third: '',
+    fourth: '',
+  },
+  expiryDate: {
+    month: '',
+    year: '',
+  },
+  ownerName: '',
+  privacyCode: '',
+  password: {
+    first: '',
+    second: '',
+    third: CRYPTO_STRING,
+    fourth: CRYPTO_STRING,
+  },
+  theme: '',
+};
+
+const CardAppPage = () => {
+  const [cardInfo, setCardInfo] = useState(initialCardInfo);
+  const [modalVisible, setModalVisible] = useState(false);
+  const [isfullFilled, setIsFullFilled] = useState(false);
+  const { number, ownerName, expiryDate, company, theme } = cardInfo;
+
+  useEffect(() => {
+    if (checkFullFilled(cardInfo)) {
+      setIsFullFilled(true);
+      return;
+    }
+
+    setIsFullFilled(false);
+  }, [cardInfo]);
+
+  const handleChange = ({ target }, item) => {
+    const { name, value } = target;
+
+    if (!validator[item](value, name)) return;
+
+    setCardInfo((prevCardInfo) => {
+      if (isObject(prevCardInfo[item])) {
+        return {
+          ...prevCardInfo,
+          [item]: {
+            ...prevCardInfo[item],
+            [name]: value,
+          },
+        };
+      }
+
+      return {
+        ...prevCardInfo,
+        [item]: value,
+      };
+    });
+  };
+
+  const handleClickCompany = (company, theme) => {
+    setCardInfo((prevCardInfo) => ({
+      ...prevCardInfo,
+      company,
+      theme,
+    }));
+
+    handleModal();
+  };
+
+  const handleModal = useCallback(() => {
+    setModalVisible((prevModalVisible) => !prevModalVisible);
+  }, []);
+
+  return (
+    <div>
+      <Header title="카드 추가">
+        <Button>
+          <PrevIcon />
+        </Button>
+      </Header>
+      <CardPreview
+        number={number}
+        ownerName={ownerName}
+        expiryDate={expiryDate}
+        company={company}
+        handleModal={handleModal}
+        theme={theme}
+      />
+      <FormInput
+        item="number"
+        inputTitle="카드 번호"
+        inputInfoList={cardNumberInputInfoList}
+        cardInfo={cardInfo}
+        onChange={handleChange}
+      />
+      <FormInput
+        className="w-50"
+        item="expiryDate"
+        inputTitle="만료일"
+        inputInfoList={expiryDateInputInfoList}
+        cardInfo={cardInfo}
+        onChange={handleChange}
+      />
+      <FormInput
+        item="ownerName"
+        inputTitle="카드 소유자 이름(선택)"
+        inputInfoList={cardOwnerNameInputInfoList}
+        cardInfo={cardInfo}
+        onChange={handleChange}
+      >
+        <div className="owner-name-length">
+          {ownerName.length} / {INPUT_MAX_LENGTH.OWNER_NAME}
+        </div>
+      </FormInput>
+      <FormInput
+        item="privacyCode"
+        inputTitle="보안코드(CVC/CVV)"
+        inputInfoList={privacyCodeInputInfoList}
+        onChange={handleChange}
+        cardInfo={cardInfo}
+      >
+        <Tooltip type="PRIVACY_CODE" />
+      </FormInput>
+      <FormInput
+        item="password"
+        inputTitle="카드 비밀번호"
+        inputInfoList={cardPasswordInputInfoList}
+        cardInfo={cardInfo}
+        onChange={handleChange}
+      />
+
+      {isfullFilled && (
+        <Button theme={theme} className="next-button">
+          다음
+        </Button>
+      )}
+      {modalVisible && (
+        <Modal handleModal={handleModal}>
+          <div className="flex-wrap">
+            {cardCompanyList.map(({ id, company, theme }) => (
+              <CardCompany
+                key={id}
+                company={company}
+                theme={theme}
+                onClickCompany={handleClickCompany}
+              />
+            ))}
+          </div>
+        </Modal>
+      )}
+    </div>
+  );
+};
+
+export default CardAppPage;
